pyproject.toml
[tool.poetry]
name = "tivit"
version = "0.1.0"
description = ""
authors = ["Michel Clemer Monteiro Deolino <michelclemermonteirodeolino@gmial.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = "^0.30.0"
alembic = "^1.13.1"
sqlmodel = "^0.0.18"
pydantic-settings = "^2.2.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
jwt = "^1.3.1"
dependency-injector = "^4.41.0"
python-jose = "^3.3.0"
psycopg2-binary = "^2.9.9"
email-validator = "^2.1.1"
pandas = "^2.2.2"
lxml = "^5.3.0"
openpyxl = "^3.1.5"
python-slugify = "^8.0.4"
redis = "^5.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
